requirements.txt
pytest~=6.2.2
Click~=7.0
Flask~=1.1.1
jsonschema~=3.2.0
MarkupSafe~=1.1.0
python-dateutil~=2.7.3